+++ apps/infrastructure/CHANGELOG.md
@@ -1,5 +1,36 @@
 # infrastructure
 
+## 0.3.2
+
+### Patch Changes
+
+- 0ba22dd: Allow fetching images from DevPortal subdomains
+
+## 0.3.1
+
+### Patch Changes
+
+- ecf2fc4: Update CDN url using the new DNS in task definition
+- 6be6508: Configure a CDN alias and add a DNS A record to ensure the DNS is accessible
+
+## 0.3.0
+
+### Minor Changes
+
+- b4f0221: Create DNS for CMS Media Library and link it to the CDN
+
+## 0.2.0
+
+### Minor Changes
+
+- 60e0be7: [DEV-1489] - Add Active Campaign records
+
+## 0.1.2
+
+### Patch Changes
+
+- fd08262: [DEV-1525] Enable auto_minor_version_upgrade to cms database
+
 ## 0.1.1
 
 ### Patch Changes